requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
openai==1.3.5
streamlit==1.28.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
tqdm==4.66.1
colorama==0.4.6
pytest-repeat==0.9.1
matplotlib==3.8.2
rich==13.7.0
jsonschema==4.20.0